Set localStorage correctly when user has no team and joins a team
This is necessary because MorScout uses the information. It would be
good if in the future, MorScout did something similar to MorTeam and
got user data from the server on each reload because this way, it is
easy for the information to get out of sync if a user's position is
updated for example.