Fix some issues with chat
If a user loaded the page as someone else starts or stops typing, an
error was thrown and the user was not able to send messages.
Additionally, after loading many messages in a chat, switching back to
that chat was very slow. This commit unloads most of the messages in a
chat before switching away.